make prewrite an action
