+++ src/storage/txn/mod.rs
@@ -9,6 +9,8 @@ pub mod scheduler;
 mod actions;
 
 pub use actions::commit::commit;
+pub use actions::prewrite::prewrite;
+pub use actions::shared::prewrite_key_value;
 
 mod latch;
 mod store;
@@ -234,4 +236,5 @@ impl ErrorCodeExt for Error {
 pub mod tests {
     use super::*;
     pub use actions::commit::tests::{must_err as must_commit_err, must_succeed as must_commit};
+    pub use actions::prewrite::tests::{try_prewrite_check_not_exists, try_prewrite_insert};
 }
